requirements.txt
Django==2.1
django-nocaptcha-recaptcha==0.0.20
django-ratelimit==1.1.0
django-webpack-loader==0.6.0
flake8==3.5.0
isort==4.3.4
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
